qa/workunits/rbd: use image id when probing for image presence
In "image resync" test we wait first for the image to be deleted
and then check that it is recreated by the mirror. Before this
fix there was a race when the image was deleted and recreated in
time window when the image presence probe was not executed. As a
result when the probe was executed later it was returning false
positive.

Signed-off-by: Mykola Golub <mgolub@mirantis.com>